refactor(capstone): extract AArch64 alias bridge into its own module
Move the Capstone->IR alias-normalization cluster (ConvertOutcome,
convert_capstone_op, the move-wide movz/movn re-encoders, and the
cinc/cinv/cneg conditional-select normalizers) out of the 8150-line
src/main.rs into a dedicated src/capstone_bridge.rs.

The new module presents a small, deep interface -- convert_capstone_op
and ConvertOutcome -- and keeps every helper private. main.rs keeps only
the thin Capstone-typed adapters (convert_capstone_op_for_optimization,
convert_to_ir) that iterate capstone::Instructions and delegate to the
bridge. The parser::parse_line delegation that guarantees the asm-text
and ELF/Capstone paths share one mnemonic set is preserved verbatim.

The convert_capstone_op tests -- including the
convert_capstone_op_handles_all_supported_aarch64_mnemonics regression
tripwire -- move with the code, and two new characterization tests pin
the wide-immediate movz re-encoding and the cinc condition inversion at
the new public seam. CLAUDE.md references updated to the new location.

No behaviour change: 1570 lib + 151 bin + 57 integration tests pass.
